// src/types/form/profile/ProfileForm.ts
export type ProfileForm = {
  name: string
  nickname: string
  career_id: string
  gender_id: string
  ageRange_id: string
  dominantHand_id: string
  playFrequency_id: string
  tennisLevel_id: string
  birthYear: string // 生年月日の年
  birthMonth: string // 生年月日の月
  birthDay: string // 生年月日の日
}